# tests/golden/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "ChampionEvaluationSnapshot": {
        "properties": {
          "artifact_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Uri"
          },
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "golden_set_version": {
            "title": "Golden Set Version",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          }
        },
        "required": [
          "commit_sha",
          "model_version",
          "golden_set_version"
        ],
        "title": "ChampionEvaluationSnapshot",
        "type": "object"
      },
      "ChampionExportEntry": {
        "properties": {
          "champion_run_id": {
            "title": "Champion Run Id",
            "type": "string"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionEvaluationSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "promoted_at": {
            "title": "Promoted At",
            "type": "string"
          },
          "promoted_by": {
            "title": "Promoted By",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "champion_run_id",
          "promoted_at",
          "promoted_by"
        ],
        "title": "ChampionExportEntry",
        "type": "object"
      },
      "ChampionExportResponse": {
        "properties": {
          "champions": {
            "items": {
              "$ref": "#/components/schemas/ChampionExportEntry"
            },
            "title": "Champions",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "ChampionExportResponse",
        "type": "object"
      },
      "ChampionHistoryEntry": {
        "properties": {
          "champion_run_id": {
            "title": "Champion Run Id",
            "type": "string"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionEvaluationSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "promoted_at": {
            "title": "Promoted At",
            "type": "string"
          },
          "promoted_by": {
            "title": "Promoted By",
            "type": "string"
          },
          "promotion_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotion Reason"
          }
        },
        "required": [
          "id",
          "model_name",
          "champion_run_id",
          "promoted_at",
          "promoted_by",
          "is_active"
        ],
        "title": "ChampionHistoryEntry",
        "type": "object"
      },
      "ChampionHistoryResponse": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChampionHistoryEntry"
            },
            "title": "History",
            "type": "array"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": [
          "model_name"
        ],
        "title": "ChampionHistoryResponse",
        "type": "object"
      },
      "ChampionResponse": {
        "properties": {
          "champion_run_id": {
            "title": "Champion Run Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "promoted_at": {
            "title": "Promoted At",
            "type": "string"
          },
          "promoted_by": {
            "title": "Promoted By",
            "type": "string"
          },
          "promotion_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotion Reason"
          }
        },
        "required": [
          "id",
          "model_name",
          "champion_run_id",
          "promoted_at",
          "promoted_by",
          "is_active"
        ],
        "title": "ChampionResponse",
        "type": "object"
      },
      "ChampionRollbackRecordResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "from_champion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Champion Id"
          },
          "from_run_id": {
            "title": "From Run Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rollback_metadata": {
            "additionalProperties": true,
            "title": "Rollback Metadata",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "to_champion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Champion Id"
          },
          "to_run_id": {
            "title": "To Run Id",
            "type": "string"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_name",
          "from_champion_id",
          "to_champion_id",
          "from_run_id",
          "to_run_id",
          "actor",
          "reason",
          "validation_status",
          "status",
          "created_at"
        ],
        "title": "ChampionRollbackRecordResponse",
        "type": "object"
      },
      "ChampionRollbackRequest": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "rolled_back_by": {
            "default": "api",
            "title": "Rolled Back By",
            "type": "string"
          },
          "target_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Run Id"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ChampionRollbackRequest",
        "type": "object"
      },
      "ChampionRollbackResponse": {
        "properties": {
          "champion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "from_champion_id": {
            "title": "From Champion Id",
            "type": "string"
          },
          "from_run_id": {
            "title": "From Run Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rolled_back_by": {
            "title": "Rolled Back By",
            "type": "string"
          },
          "to_run_id": {
            "title": "To Run Id",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "from_champion_id",
          "from_run_id",
          "to_run_id",
          "rolled_back_by",
          "reason",
          "dry_run"
        ],
        "title": "ChampionRollbackResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromoteChampionRequest": {
        "properties": {
          "promoted_by": {
            "default": "api",
            "title": "Promoted By",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "PromoteChampionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ares",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/champions/export": {
      "get": {
        "operationId": "export_api_v1_champions_export_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}": {
      "get": {
        "operationId": "get_champion_api_v1_champions__model_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Champion",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}/history": {
      "get": {
        "operationId": "history_api_v1_champions__model_name__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}/previous": {
      "get": {
        "operationId": "previous_api_v1_champions__model_name__previous_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Previous",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}/promote": {
      "post": {
        "operationId": "promote_api_v1_champions__model_name__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteChampionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}/rollback": {
      "post": {
        "operationId": "rollback_api_v1_champions__model_name__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionRollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionRollbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback",
        "tags": [
          "champions"
        ]
      }
    },
    "/api/v1/champions/{model_name}/rollbacks": {
      "get": {
        "operationId": "rollbacks_api_v1_champions__model_name__rollbacks_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChampionRollbackRecordResponse"
                  },
                  "title": "Response Rollbacks Api V1 Champions  Model Name  Rollbacks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollbacks",
        "tags": [
          "champions"
        ]
      }
    },
...
  }
}